Add files via upload
Utfør fikser i hennhold til todo.txt 
Implementert brukersystem og mulighet for å endre enkelte IC-verdier over CANbus, ref:  ICmodificationComands.txt

+++ AR19_LC/MainBoard/src/errorHandler.h
@@ -24,8 +24,8 @@ class ErrorHandler
     
     private:
 
-    uint16_t numerOfErrors;
-    uint16_t _componentID;
+    uint8_t numerOfErrors;
+    uint8_t _componentID;
     
     errorMessage _errorLog[10];
